Finished VMapFactory, yet to test til servers are further in development.

+++ WrathForged.Common/Scripting/Interfaces/CoreEvents/IUpdateLoop.cs
@@ -14,5 +14,5 @@ namespace WrathForged.Common.Scripting.Interfaces.CoreEvents;
 /// </summary>
 public interface IUpdateLoop
 {
-    void UpdateTick(uint diff);
+    void UpdateTick(double diff);
 }
